+++ cmd/llpkgstore/main.go
@@ -1,0 +1,7 @@
+package main
+
+import cmd "github.com/goplus/llpkgstore/cmd/llpkgstore/internal"
+
+func main() {
+	cmd.Execute()
+}
