+++ config/config.go
@@ -3,8 +3,8 @@ package config
 import (
 	"errors"
 
-	"github.com/goplus/llpkgstore/tools/pkg/upstream"
-	"github.com/goplus/llpkgstore/tools/pkg/upstream/installer/conan"
+	"github.com/goplus/llpkgstore/upstream"
+	"github.com/goplus/llpkgstore/upstream/installer/conan"
 )
 
 var ValidInstallers = []string{"conan"}
