+++ config/validate_test.go
@@ -3,7 +3,7 @@ package config
 import "testing"
 
 func TestValidateLLpkgConfig(t *testing.T) {
-	config, err := ParseLLpkgConfig("../../_demo/llpkg.cfg")
+	config, err := ParseLLpkgConfig("../_demo/llpkg.cfg")
 	if err != nil {
 		t.Errorf("Error parsing config file: %v", err)
 	}
