+++ go.mod
@@ -1,3 +1,10 @@
 module github.com/goplus/llpkgstore
 
-go 1.24.0
+go 1.23
+
+require github.com/spf13/cobra v1.9.1
+
+require (
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/spf13/pflag v1.0.6 // indirect
+)
