--- tools/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/goplus/llpkgstore/tools/cmd"
-
-func main() {
-	cmd.Execute()
-}
