+++ upstream/installer/conan/conaninst_test.go
@@ -7,7 +7,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/goplus/llpkgstore/tools/pkg/upstream"
+	"github.com/goplus/llpkgstore/upstream"
 )
 
 func TestConanInstaller(t *testing.T) {
